# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OCC Operational Control Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: ptId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskWithRelations"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithRelations"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithRelations"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Cancel task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithRelations"
  /tasks/{id}/comments:
    get:
      operationId: listTaskComments
      tags: [tasks]
      summary: List task comments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [tasks]
      summary: Add task comment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"
...
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [systemSettings]
      summary: List audit logs
      security:
        - bearerAuth: []
      parameters:
        - name: module
          in: query
          schema:
            type: string
        - name: actionType
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /inactivity/check:
    get:
      operationId: checkInactivity
      tags: [systemSettings]
      summary: Check dealer inactivity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Inactivity report
          content:
            application/json:
              schema:
                type: object
  /system-settings:
    get:
      operationId: listSystemSettings
      tags: [systemSettings]
      summary: List system settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemSetting"
  /system-settings/{key}:
    put:
      operationId: updateSystemSetting
      tags: [systemSettings]
      summary: Update system setting
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSystemSettingRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSetting"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    TaskWithRelations:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        assignedTo:
          type: integer
        assigneeName:
          type: string
        assignedBy:
          type: integer
        assignerName:
          type: string
        priority:
          type: string
        deadline:
          type: string
          format: date-time
        progressPercent:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, priority, progressPercent, status]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        assignedTo:
          type: integer
        priority:
          type: string
        deadline:
          type: string
          format: date-time
      required: [title]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: integer
        priority:
          type: string
        deadline:
          type: string
          format: date-time
        progressPercent:
          type: integer
        status:
          type: string
    TaskComment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, taskId, userId, message]
    CreateTaskCommentRequest:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SystemSetting:
      type: object
      properties:
        id:
          type: integer
        settingKey:
          type: string
        settingValue:
          type: string
        description:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [id, settingKey]
    UpdateSystemSettingRequest:
      type: object
      properties:
        settingValue:
          type: string
        description:
          type: string
      required: [settingValue]